提交说明： 1.添加敌人 2.敌人信息部分完善 3.关卡介绍信息完善 4.调整敌人速度

+++ Assets/Editor/TowerInfoEditor.cs
@@ -38,7 +38,7 @@ public class TowerInfoEditor:BaseExcelEditor
                     info.sellCoin = GetInt(4);
                     info.damageType = GetInt(5);
                     info.damageRange = GetFloat(6);
-                    info.CD = GetInt(7);
+                    info.CD = GetFloat(7);
                     info.damage = GetInt(8);
                     info.Range = GetInt(9);
                     info.bullectPath = GetString(10);
